[REFACTOR] Phase out legacy go ffi
This PR phases out te legacy go FFI as so we can focus on the new FFI, and phase out legacy C APIs.
The change will allow us to move away from the legacy FFI API and simplify the project maintainace.
A new bindings can be built on top of new FFI later if there is community demand.