[FFI] Serialization To/From JSONGraph
This PR implements serialization function for generic ffi::Any
based on the reflection that is preserves the overall object
graph reference relation. These extra APIs are implemented
through the reflection system.

They can be used to further modernize and unify the serialization
mechanisms in the project under the new reflection mechanism.