[Fix] Update FlashInfer JIT header lookup (#18244)
This PR fixes the tvm/dlpack/dmlc header lookup in the FlashInfer
kernel JIT compilation.

Prior to this fix, the JIT compilation assumes the environment
variable `TVM_SOURCE_DIR` is always defined, which is not always
true. This PR fixes the behavior and considers multiple cases,
including TVM source builds and pip-installed packages.