[CUDA] Support NVTX in CUDA 13 (#18300)
This PR adds the support of NVTX for CUDA 13. The change is because
that starting CUDA 13, the nvtx functions are moved to the lirbary
of `libnvtx3interop.so`, and the previous nvToolsExt library no
longer exists.

To ensure compatibility with both CUDA 12 and 13, we add
libnvtx3interop.so to the library lookup list.