[Bugfix] Prevent segfault when instantiating abstract SearchStrategy (#18534)
Add __init__ method to SearchStrategy class to prevent direct
instantiation of the abstract class. This raises a TypeError with a
helpful error message instead of causing a segmentation fault when
SearchStrategy() is called directly or passed to TuneContext.

Also add additional check in TuneContext.__init__ to ensure abstract
SearchStrategy instances are not used.

Fixes #18268